Added support for visiting elements in vectors
Defined new macro DEFINE_MEMBER_VECTOR to enable vising of each member
of vector. Because ther are types that are iterable,
but we still want to treat them as a single value (i.e std::string),
the information whether member shall be treated as vector or a single
element must come from member type definition itself.